broom/parser.py
"""
Parser module for handling Go code parsing.
"""

from pathlib import Path
from typing import Any, Dict, List

import tree_sitter


class GoParser:
    """Class for parsing Go code."""

    def __init__(self):
        """Initialize the Go parser."""
        # Note: In a real implementation, you would need to build the Go grammar
        # and load it here. This is a placeholder for the actual implementation.
        self.parser = tree_sitter.Parser()
        # self.parser.set_language(
        #     Language('build/my-languages.so', 'go')
        # )

    def parse_file(self, file_path: Path) -> List[Dict[str, Any]]:
        """Parse a Go file and extract code blocks.

        Args:
            file_path: Path to the Go file

        Returns:
            List of code blocks with metadata
        """
        # This is a placeholder implementation
        # In a real implementation, you would:
        # 1. Parse the file using tree-sitter
        # 2. Extract functions, methods, and other code blocks
        # 3. Return them with appropriate metadata
        content = file_path.read_text(encoding="utf-8")
        return [{
            "code": content,
            "file_path": str(file_path),
            "type": "file"
        }]

    def parse_repository(self, repo_path: Path) -> List[Dict[str, Any]]:
        """Parse all Go files in a repository.

        Args:
            repo_path: Path to the repository

        Returns:
            List of code blocks with metadata
        """
        go_files = list(repo_path.rglob("*.go"))
        all_blocks = []

        for file_path in go_files:
            if not any(part.startswith(".") for part in file_path.parts):
                blocks = self.parse_file(file_path)
                all_blocks.extend(blocks)

        return all_blocks
