requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
mccabe==0.7.0
mypy_extensions==1.1.0
orjson==3.10.18
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.11.1
pydantic==2.11.6
pydantic_core==2.33.2
pyflakes==3.2.0
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
torch==2.2.1
transformers==4.38.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
zstandard==0.23.0
